+++ rick-and-morty/src/components/AntdTable/AntdTable.tsx
@@ -1,15 +1,9 @@
 import { Button, Table } from "antd";
-import type { ColumnsType } from "antd/es/table";
 import React, { useEffect, useState } from "react";
 import { CharactersService } from "../../services/characters.service";
+import { columns } from "./columns";
 
-export interface DataType {
-  key: string;
-  name: string;
-  status: string;
-  species: string;
-  gender: string;
-}
+
 
 const TestTable: React.FC = () => {
   const [totalPages, setTotalPages] = useState<number | undefined>(undefined);
@@ -18,12 +12,12 @@ const TestTable: React.FC = () => {
   useEffect(() => {
     const fetchData = async () => {
       try {
-        const response = await CharactersService.getCharactees(curPage);
-        if (response.data.info.next === null) {
-          setCurPage(response.data.info.pages);
+        const { data } = await CharactersService.getCharactees(curPage);
+        if (data.info.next === null) {
+          setCurPage(data.info.pages);
         } else {
-          setTableData(response.data.results);
-          setTotalPages(response.data.info.pages);
+          setTableData(data.results);
+          setTotalPages(data.info.pages);
         }
       } catch (error) {
         console.error("Ошибка в получении данных", error);
@@ -45,28 +39,6 @@ const TestTable: React.FC = () => {
     }
   };
 
-  const columns: ColumnsType<DataType> = [
-    {
-      title: "Имя",
-      dataIndex: "name",
-      key: "name",
-    },
-    {
-      title: "Статус",
-      dataIndex: "status",
-      key: "status",
-    },
-    {
-      title: "Вид",
-      dataIndex: "species",
-      key: "species",
-    },
-    {
-      title: "Гендер",
-      key: "gender",
-      dataIndex: "gender",
-    },
-  ];
 
   return (
     <>
@@ -81,7 +53,7 @@ const TestTable: React.FC = () => {
         &gt;
       </Button>
     </>
-  )
+  );
 };
 
 export default TestTable;
